otel-demo-stack/api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-exporter-otlp-proto-http>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
